# openapi-spec.yaml
openapi: "3.0.3"
info:
  version: 0.0.1
  title: Formel 1 API
  description: A small sample API to grasp the basics of the OpenAPI Specification
servers:
  - url: http://localhost:8080/api/v1/f1
paths:
  /driver:
    get:
      summary: List all drivers (funktioniert noch nicht)
      parameters:
        - name: limit
          in: query
          description: How many items to return at one time
          required: false
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: An paged array of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriverDto"
  /racingteam:
    get:
      summary: List all Racing Teams (funktioniert noch nicht)
      parameters:
        - name: limit
          in: query
          description: How many items to return at one time
          required: false
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: An paged array of Racing Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RacingTeamDto"
  /ponies:
    get:
      summary: List all ponies
      operationId: getPonies
      parameters:
        - name: limit
          in: query
          description: How many items to return at one time
          required: false
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: An paged array of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PonyDto"
    post:
      summary: introduce a new pony to the flock
      operationId: addPony
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PonyDto"
      responses:
        201:
          description: added the pony
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PonyDto"
components:
  schemas:
    DriverDto:
      required:
        - name
        - prename
        - age
        - country
        - points
      properties:
        name:
          type: string
          example: Müller
          minLength: 1
          maxLength: 255
        prename:
          type: string
          example: Hans
          minLength: 1
          maxLength: 255
        age:
          type: integer
          example: 20
          minimum: 18
          maximum: 100
        racingTeam:
          type: object
        points:
          type: integer
          example: 15
          minimum: 0
          maximum: 600
        country:
          type: string
          example: Switzerland
          minLength: 1
          maxLength: 255
    RacingTeamDto:
      required:
        - name
        - teamBoss
        - engineProducer
        - country
        - points
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        teamBoss:
          type: string
          minLength: 1
          maxLength: 255
        engineProducer:
          type: string
          minLength: 1
          maxLength: 255
        country:
          type: string
          minLength: 1
          maxLength: 255
        points:
          type: integer
          minimum: 0
          maximum: 1200
    PonyDto:
      required:
        - name
        - birthday
      properties:
        id:
          type: integer
          example: 42
          minimum: 0
          readOnly: true
        name:
          type: string
          minLength: 1
          example: Thunder
        birthday:
          type: string
          format: date
          example: 2021-01-31